requirements.txt
streamlit>=1.36.0
pymupdf>=1.24.0
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
numpy>=1.26.0
requests>=2.32.0
tqdm>=4.66.0
